mise à jour des routes, ajout de l'admin controller

+++ app/Http/Controllers/BudgetController.php
@@ -2,9 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Budget;
+use Exception;
 use Illuminate\Http\Request;
+use function PHPSTORM_META\type;
+
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 
 class BudgetController extends Controller
 {
@@ -15,7 +20,7 @@ class BudgetController extends Controller
      */
     public function index()
     {
-        return response()->json(Budget::all());
+        return response()->json(Auth::user()->budgets()->get());
     }
 
     /**
@@ -29,17 +34,17 @@ class BudgetController extends Controller
         $request->validate([
             'value' => 'required',
             'field_id' => 'required',
-            'user_id' => 'required',
             'month_id' => 'required',
             'year_id' => 'required',
         ]);
+
+        $request['user_id'] = Auth::id();
 
         Budget::create($request->all());
 
         return response([
             'message' => "Création de la valeur Budget {$request->value} réussie",
             'donnees' => Budget::query()->orderBy('id', 'desc')->first(),
-            //'id' => Budget::query()->orderBy('id', 'desc')->first()->id,
         ]);
     }
 
@@ -51,7 +56,20 @@ class BudgetController extends Controller
      */
     public function show(Budget $budget)
     {
-        return response()->json(Budget::find($budget->id));
+        $result = Budget::where([
+            ['id', '=', $budget->id],
+            ['user_id', '=', Auth::id()]
+        ])->get();
+        
+        if(Auth::user()->status->name == 'Admin'){
+            return response()->json(Budget::find($budget->id));
+        }else{
+            if (isset($result[0])) {
+                return response()->json($result);
+            } else {
+                return ['message' => "La valeur de ce budget n'est pas accessible avec ce profil"];
+            }
+        }
     }
 
     /**
@@ -63,13 +81,19 @@ class BudgetController extends Controller
      */
     public function update(Request $request, Budget $budget)
     {
-        $budget = Budget::findOrFail($budget->id);
-        $budget->update($request->all());
-        return response([
-            'message' => "Modification de la valeur Budget id {$budget->id} réussie",
-            'donnees' => $budget
-
-        ]);
+        if(Auth::user()->id == $budget->user_id){
+            $budget = Budget::findOrFail($budget->id);
+            $budget->update($request->all());
+            return response([
+                'message' => "Modification de la valeur Budget id {$budget->id} réussie",
+                'donnees' => $budget
+            ], 201);
+        }
+        else{
+            return response([
+                'message' => "Modification de la valeur Budget impossible, droits insuffisants"
+            ], 401);
+        }
     }
 
     /**
@@ -80,11 +104,16 @@ class BudgetController extends Controller
      */
     public function destroy(Budget $budget)
     {
-        Budget::findOrFail($budget->id)->delete();
-        return response([
-            'message' => "Suppression de la valeur Budget id {$budget->id} réussie",
-            'donnees' => $budget
-
-        ]);
+        if (Auth::user()->id == $budget->user_id) {
+            Budget::findOrFail($budget->id)->delete();
+            return response([
+                'message' => "Suppression de la valeur Budget id {$budget->id} réussie",
+                'donnees' => $budget
+            ]);
+        } else {
+            return response([
+                'message' => "Suppression de la valeur Budget impossible, droits insuffisants"
+            ], 401);
+        }
     }
 }
